docs: update metrics docs (semgrep/semgrep-proprietary#3960)
## What

Updated documentation for the CLI regarding when metrics is set to auto.
Before, documentation stated that metrics will not be sent when only
using configs from local files. It was changed to state that when using
local files, metrics will be sent when the user is logged in. This was
updated in several locations throughout the code base where relevant.

## Why

Documentation did not match intended behavior. Addresses issue #11028.  

## Test plan

Recompiled with `make dev` and ran `semgrep scan --help` to verify that
the documentation referred by the GitHub issue was correctly updated.

Closes https://github.com/semgrep/semgrep/issues/11028.

synced from Pro 964bd91540a44f21fe96ebccac54a469c0daf730