feat(tracing): switch to otel only tracing (semgrep/semgrep-proprietary#3971)
This PR gets rid of our dependency on Otel Trace, which was a backend for the ocaml tracing library, and switches us to using the otel only implementation. This let's us refactor a bunch of code/cleanup easily, and allows us to use some more otel specific features for tracing, such as events, span kind, and some easier linking with metrics.

## Test plan
We don't have much in the way of testing for traces, but I did some manual testing:

https://app.datadoghq.com/apm/trace/db6bd649744d9e8b3699155e86dc11db?graphType=waterfall&shouldShowLegend=true&spanID=3456133268998997357&spanViewType=metadata
https://app.datadoghq.com/apm/trace/c6a592f0af94b0d6102961913a960c48?graphType=waterfall&shouldShowLegend=true&spanID=5725073337248330188&spanViewType=metadata

synced from Pro 06e100f166d5037de903048810692945803aece5