feat(js/ts): update to latest (semgrep/semgrep-proprietary#4068)
## What:
This PR updates our JS/TS parser by quite a bit.

## Why:
Originally, we set out to kill our Menhir parser. However, on doing
this, we had to rip out the JSON parser, which was really the Typescript
parser in disguise. Upon doing this ,we observed that rule loading times
went from 10s to 30s, a significant regression.

Instead of doing that, in the meantime we will just land these parser
changes, so we can stop blocking our `ocaml-tree-sitter-core`
development pipeline.

## How:
The majority of this work, as much as it desired to completely remove
the necrotizing influence of Menhir, could not escape it completely. The
`pfff` parser enabled a wide range of edge behaviors for pattern
parsing, which we could not justifiably kill in its entirety. Parts of
it which I deemed to be obsolete or rare were not implemented, but the
majority of edge parsing cases that `pfff` implemented, we
re-implemented in `ocaml-tree-sitter-semgrep`.

These are the behaviors that we decided not to support:
- Partial class patterns, such as `class $X`. These can be written as
`class $X { ... }`.
- Ellipses in XML attribute value position, such as `<foo href=.../>`.
These can be written as `<foo href=$_/>`
- Partial method definitions, such as 
```
@Query
async foo
```
and
```
@Query
async foo(...)
```
These are preferred as 
```
@Query
async foo(...) { ... }
```

(Note that the behaviors we do not support do not matter in this PR in
particular, as we still have the Menhir parser around)

Otherwise, we successfully reimplemented all `pfff` Semgrep pattern
behavior, such as metavariable ellipses, dot access chaining, ellipses
in particular places, deep ellipses, and a bug where automatic
semicolons were not inserted before ellipses, among others.

## Test plan:
Automated tests, and parsing rate:
```
{
  "name": "*",
  "parsing_rate": 0.9996404734174135,
  "line_count": 5562871,
  "error_line_count": 2000,
  "file_count": 37247,
  "error_file_count": 572,
  "total_node_count": 60131659,
  "untranslated_node_count": 197688
}
```

synced from Pro c9c18c7ab23f6e1e9b23f8b14ce0aa033c081c19