fix(sca): include `allowed_hashes` in pnpm lockfile parser (semgrep/semgrep-proprietary#4404)
To update my pnpm code for #4397 I reorganized the logic a bit and
started parsing the sha512 hash for each dependency from
`pnpm-lock.yaml`. Based on our example lockfiles, the hash has the same
location and format in every version of the pnpm lockfile format that we
support, so this should work for all lockfiles.

If we can't parse a hash for whatever reason, we just skip it. I'm not
sure whether we use the `allowed_hashes` field for anything yet, so this
seemed like the best default.

Test plan: tested change locally

synced from Pro 60870bd53e806e11310acb9435bf8005af7a67e1